requirements.txt
flask==2.3.3
flask-sqlalchemy==2.5.1
flask-cors==4.0.0
flask-limiter==3.5.0
gunicorn==21.2.0
dnspython==2.4.2
email-validator==2.1.0
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
werkzeug==2.3.7
sqlalchemy==1.4.53
